refactor: organise tests according to dependencies (#257)
* refactor: organise tests according to dependencies

- organises tests according to their dependencies.
- removes unnecessary autouse fixtures
- removes gated imports at the module level
- uses `--ignore` to exclude redis/sqlalchemy/sentry tests from no-extras tests

* refactor: use `test_ignore_glob` pytest global

This lets us skip test collection for a whole sub-package of tests if
a dependency is unavailable.

* refactor: log tests can run without sqlalchemy.

This was only necessary due to the dependency on
`tests.utils.controllers` in the `http_scope` fixture, which has now
been removed.

* refactor: don't need to skipif saq log tests

As the log tests depend on `job` fixture which will automatically skip
if SAQ not available.

* refactor: remaining fixtures/tests requiring sqlalchemy moved into sqlalchemy directory.
