+++ mypy.ini
@@ -18,9 +18,9 @@ warn_unused_ignores = True
 [mypy-tests.*]
 disallow_untyped_decorators = False
 
-[mypy-tests.unit.test_dto]
+[mypy-tests.unit.sqlalchemy.test_dto]
 # for the declarative base fixture
 disable_error_code = valid-type,misc
 
-[mypy-tests.unit.repository.test_sqlalchemy]
+[mypy-tests.unit.sqlalchemy.repository.test_sqlalchemy]
 disable_error_code = attr-defined
