+++ tests/unit/conftest.py
@@ -5,6 +5,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 import pytest
+from starlite import get
 from starlite.datastructures import State
 from starlite.enums import ScopeType
 
@@ -15,66 +16,6 @@ if TYPE_CHECKING:
     from saq.job import Job
     from starlite import Starlite
     from starlite.types import HTTPResponseBodyEvent, HTTPResponseStartEvent, HTTPScope
-
-    from starlite_saqlalchemy.testing.generic_mock_repository import (
-        GenericMockRepository,
-    )
-    from tests.utils.domain.authors import Author
-    from tests.utils.domain.books import Book
-
-
-@pytest.fixture(name="author_repository_type", autouse=constants.IS_SQLALCHEMY_INSTALLED)
-def fx_author_repository_type(
-    authors: list[Author], monkeypatch: pytest.MonkeyPatch
-) -> type[GenericMockRepository[Author]]:
-    """Mock Author repository, pre-seeded with collection data."""
-    from starlite_saqlalchemy.testing.generic_mock_repository import (
-        GenericMockRepository,
-    )
-    from tests.utils.domain.authors import Author
-    from tests.utils.domain.authors import Service as AuthorService
-
-    repo = GenericMockRepository[Author]
-    repo.seed_collection(authors)
-    monkeypatch.setattr(AuthorService, "repository_type", repo)
-    return repo
-
-
-@pytest.fixture(name="author_repository", autouse=constants.IS_SQLALCHEMY_INSTALLED)
-def fx_author_repository(
-    author_repository_type: type[GenericMockRepository[Author]],
-) -> GenericMockRepository[Author]:
-    """Mock Author repository instance."""
-    return author_repository_type()
-
-
-@pytest.fixture(name="book_repository_type", autouse=constants.IS_SQLALCHEMY_INSTALLED)
-def fx_book_repository_type(
-    books: list[Book], monkeypatch: pytest.MonkeyPatch
-) -> type[GenericMockRepository[Book]]:
-    """Mock Book repository, pre-seeded with collection data."""
-    from starlite_saqlalchemy.testing.generic_mock_repository import (
-        GenericMockRepository,
-    )
-    from tests.utils.domain.books import Book
-    from tests.utils.domain.books import Service as BookService
-
-    class BookRepository(GenericMockRepository[Book]):
-        """Mock book repo."""
-
-        model_type = Book
-
-    BookRepository.seed_collection(books)
-    monkeypatch.setattr(BookService, "repository_type", BookRepository)
-    return BookRepository
-
-
-@pytest.fixture(name="book_repository", autouse=constants.IS_SQLALCHEMY_INSTALLED)
-def fx_book_repository(
-    book_repository_type: type[GenericMockRepository[Book]],
-) -> GenericMockRepository[Book]:
-    """Mock Book repo instance."""
-    return book_repository_type()
 
 
 @pytest.fixture()
@@ -96,7 +37,10 @@ def http_response_body() -> HTTPResponseBodyEvent:
 @pytest.fixture()
 def http_scope(app: Starlite) -> HTTPScope:
     """Minimal ASGI HTTP connection scope."""
-    from ..utils import controllers
+
+    @get()
+    def handler() -> None:
+        ...
 
     return {
         "headers": [],
@@ -111,7 +55,7 @@ def http_scope(app: Starlite) -> HTTPScope:
         "query_string": b"",
         "raw_path": b"/wherever",
         "root_path": "/",
-        "route_handler": controllers.get_author,
+        "route_handler": handler,
         "scheme": "http",
         "server": None,
         "session": {},
