+++ tests/unit/redis/worker/conftest.py
@@ -1,0 +1,4 @@
+from starlite_saqlalchemy.constants import IS_SAQ_INSTALLED
+
+if not IS_SAQ_INSTALLED:
+    collect_ignore_glob = ["*"]
