+++ tests/unit/sentry/conftest.py
@@ -1,0 +1,4 @@
+from starlite_saqlalchemy.constants import IS_SENTRY_SDK_INSTALLED
+
+if not IS_SENTRY_SDK_INSTALLED:
+    collect_ignore_glob = ["*"]
